added question before deleting, added update (exept rent)

+++ ToDoApp/FrmUpdate.cs
@@ -43,7 +43,17 @@ namespace ToDoApp
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            DBCars.DeleteCar(car.Id);
+            if(MessageBox.Show("Are you sure that you want to delete " + car.Company + " - " + car.Model +"?","Info", MessageBoxButtons.YesNo) == DialogResult.Yes)
+            {
+                DBCars.DeleteCar(car.Id);
+                closeAndRefresh();
+            }
+            
+        }
+
+        private void btnSave_Click(object sender, EventArgs e)
+        {
+            DBCars.UpdateCar(car.Id,  cboCompanyUpdate.Text, txtModelUpdate.Text, Int32.Parse(txtYearUpdate.Text));
             closeAndRefresh();
         }
     }
